Run the migration as part of the setup command
